refactor request util
