+++ packages/waku/src/lib/middleware/types.ts
@@ -1,6 +1,6 @@
 import type { Config } from '../../config.js';
 
-import type { EntriesDev, EntriesPrd } from '../types.js';
+import type { EntriesPrd } from '../types.js';
 
 export type ClonableModuleNode = { url: string; file: string };
 
@@ -8,22 +8,6 @@ export type HandlerContext = {
   req: Request;
   res?: Response;
   readonly data: Record<string, unknown>;
-  unstable_devServer?: {
-    rootDir: string;
-    resolveClientEntry: (id: string) => string;
-    loadServerModuleRsc: (idOrFileURL: string) => Promise<Record<string, any>>;
-    loadEntriesDev: (config: { srcDir: string }) => Promise<EntriesDev>;
-    loadServerModuleMain: (idOrFileURL: string) => Promise<Record<string, any>>;
-    transformIndexHtml: (
-      pathname: string,
-    ) => Promise<TransformStream<any, any>>;
-  };
-  unstable_modules?: {
-    rsdwServer: unknown;
-    rdServer: unknown;
-    rsdwClient: unknown;
-    wakuMinimalClient: unknown;
-  };
 };
 
 export type Handler = (
